+++ packages/contracts/source/contracts/p2p/broadcaster.ts
@@ -1,5 +1,6 @@
-import { ITransaction } from "../crypto";
+import { IBlock, ITransaction } from "../crypto";
 
 export interface Broadcaster {
 	broadcastTransactions(transactions: ITransaction[]): Promise<void>;
+	broadcastBlock(block: IBlock): Promise<void>;
 }
